refact: tag selector into its own component from tag search
